SVN sync
git-svn-id: http://svn.code.sf.net/p/megui/code@2233 a21b8521-4df1-4ed3-b48c-cff5bc73dcb3

+++ megui/trunk/core/util/FileUtil.cs
@@ -366,27 +366,54 @@ namespace MeGUI.core.util
         /// in job.FilesToDelete if settings.DeleteIntermediateFiles is checked
         /// </summary>
         /// <param name="job">the job which should just have been completed</param>
-        public static LogItem DeleteIntermediateFiles(List<string> files, bool bAlwaysAddLog)
+        public static LogItem DeleteIntermediateFiles(List<string> files, bool bAlwaysAddLog, bool askAboutDelete)
         {
             bool bShowLog = false;
             LogItem i = new LogItem("Deleting intermediate files");
-            
-            // delete all files first
+
+            List<string> arrFiles = new List<string>();
             foreach (string file in files)
             {
-                try
+                if (Directory.Exists(file))
+                    continue;
+                else if (!File.Exists(file))
+                    continue;
+                if (!arrFiles.Contains(file))
+                    arrFiles.Add(file);
+            }
+
+            if (arrFiles.Count > 0)
+            {
+                bShowLog = true;
+                bool delete = true;
+
+                if (askAboutDelete)
+                    delete = MainForm.Instance.DialogManager.DeleteIntermediateFiles(arrFiles);
+                if (!delete)
+                    return null;
+
+                // delete all files first
+                foreach (string file in arrFiles)
                 {
-                    if (Directory.Exists(file))
-                        continue;
-                    else if (!File.Exists(file))
-                        continue;
-                    bShowLog = true;
-                    File.Delete(file);
-                    i.LogEvent("Successfully deleted " + file);
-                }
-                catch (IOException e)
-                {
-                    i.LogValue("Error deleting " + file, e, ImageType.Error);
+                    int iCounter = 0;
+                    while (File.Exists(file))
+                    {
+                        try
+                        {
+                            File.Delete(file);
+                            i.LogEvent("Successfully deleted " + file);
+                        }
+                        catch (IOException e)
+                        {
+                            if (++iCounter >= 3)
+                            {
+                                i.LogValue("Problem deleting " + file, e.Message, ImageType.Warning);
+                                break;
+                            }
+                            else
+                                System.Threading.Thread.Sleep(2000);
+                        }
+                    }
                 }
             }
 
@@ -404,12 +431,12 @@ namespace MeGUI.core.util
                             i.LogEvent("Successfully deleted directory " + file);
                         }
                         else
-                            i.LogEvent("Did not delete " + file + " as the directory is not empty.");
+                            i.LogEvent("Did not delete " + file + " as the directory is not empty.", ImageType.Warning);
                     }  
                 }
                 catch (IOException e)
                 {
-                    i.LogValue("Error deleting directory " + file, e, ImageType.Error);
+                    i.LogValue("Problem deleting directory " + file, e.Message, ImageType.Warning);
                 }
             }
             if (bAlwaysAddLog || bShowLog)
